* chore: upgrade Java 11 to Java 17

- Update maven.compiler.source/target/release to 17 across all 64 pom.xml files
- Migrate javax.* imports to jakarta.* (servlet, annotation, persistence, ws.rs, ejb)
- Upgrade Spring Framework 5.x → 6.0.9
- Upgrade Spring Boot 2.x → 3.1.0
- Upgrade Hibernate ORM 5.x → 6.2.0.Final
- Upgrade Hibernate Search 6.1.4 → 6.2.0.Final
- Upgrade Spring Batch 4.x → 5.0.2
- Upgrade Spring Retry 1.2.2 → 2.0.1
- Upgrade Thymeleaf 3.0.14 → 3.1.1 (thymeleaf-spring5 → thymeleaf-spring6)
- Upgrade Swagger 2.1.12 → 2.2.9 (Jakarta EE support)
- Upgrade Kotlin 1.5.31 → 1.8.22 (Spring 6 compatibility)
- Fix Spring Batch 5 API: ItemWriter.write(Chunk), JobParameter(value, Class)
- Fix Hibernate 6 API: removed PostgreSQL94Dialect, Search.session() unwrap
- Fix Spring 6 API: removed WebMvcConfigurerAdapter, @Required, @Bean(autowire)
- Disable hibernate54-ddl-maven-plugin (incompatible with Hibernate 6)
- Update Velocity template for Jakarta servlet imports


* fix: address review feedback - remove incorrect @Autowired replacements

- Remove @Autowired from setResourceType, setServers, setDao (was @Required,
  which is validation-only; replacing with @Autowired changes DI semantics)
- Fix Javadoc: @Required in docs referred to @RequiredParam, not Spring's @Required
- Fix NonPersistedBatchConfigurer: add dedicated DataSourceTransactionManager
  for batch H2 datasource instead of reusing HAPI JPA transaction manager


---------

+++ hapi-fhir-test-utilities/src/main/java/ca/uhn/fhir/test/utilities/BatchJobHelper.java
@@ -143,12 +143,12 @@ public class BatchJobHelper {
 			));
 	}
 
-	public int getReadCount(Long theJobExecutionId) {
+	public long getReadCount(Long theJobExecutionId) {
 		StepExecution stepExecution = getStepExecution(theJobExecutionId);
 		return stepExecution.getReadCount();
 	}
 
-	public int getWriteCount(Long theJobExecutionId) {
+	public long getWriteCount(Long theJobExecutionId) {
 		StepExecution stepExecution = getStepExecution(theJobExecutionId);
 		return stepExecution.getWriteCount();
 	}
